requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
pydantic>=2.0.0
databricks-sdk>=0.30.0
databricks-sql-connector>=3.0.0
python-multipart>=0.0.9
mlflow>=2.18.0
